⚡ Bolt: Optimize popup render and batch storage I/O
Optimized `App.tsx` by memoizing chart calculations and batching `localStorage` operations for color assignments.

- Extracted color utilities to `extension/src/ui/colors.ts`
- Added unit tests for color utilities
- Replaced O(N) localStorage I/O with O(1) by batching
- Memoized `chartSegments` to prevent re-renders on hover
- Moved static assets outside component
